apply second round of Gygabrain feedback
- SubcategoryFieldType.from_type_int → from_type_code
- Remove "Multilingual" from TEXT/TEXTAREA docstrings
- FieldCondition.values: list[str] → set[str]; add is_satisfied_by(value) method
- SubcategoryStructure.lower_label_map: dataclass field → cached_property
- Move _parse_title_fields + _TITLE_SUFFIX_TYPES from subcategory_structure.py to parsers/utils.py
- Add OfferFields.subcategory_structure property (calls SubcategoryStructure.from_offer_fields(self))
- SubcategoryStructure docstring: "from a parsed OfferFields" → "from OfferFields"
